requirements.txt
SpeechRecognition
browserforge
bs4
requests
pytest-playwright
playwright-stealth
setuptools